+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add new prometheus metric `states{type="raid_lock"}`
 - The `LOOT_MODULE` now keeps a full history of what was rolled when, and who won what on which roll. You can search this history by using `!loot history`, `!loot history <number>|last`, `!loot search winner=Nady` and `!loot search item=leg`.
 - All Docker images now support setting fixed settings via setting environment variables `CONFIG_SETTING_<setting>=<value>`, e.g. `CONFIG_SETTING_CONSOLE_COLOR=1`
+- New commands `!config setting <name>` and `!config setting <name> admin <access level>` to change the required access level to change a setting's value.
 
 ### Changed